fix: svd to pku environment
