+++ demo-dag-svd-python/code/stage0/requirements.txt
@@ -1,0 +1,1 @@
+numpy==2.1.3
